protect endpoints

+++ backend/routers/todos.js
@@ -1,22 +1,23 @@
 const { Todo } = require('../models/todo');
-// const auth = require('../middleware/auth');
+const auth = require('../middleware/auth');
 const express = require('express');
 const Joi = require('joi');
 
 const router = express.Router();
 
-router.get('/', async (req, res) => {
+router.get('/', auth, async (req, res) => {
   try {
     const todos = await Todo.find().sort({ date: -1 });
+    const filterdTodos = todos.filter((todo) => todo.uid === req.user._id);
 
-    res.send(todos);
+    res.send(filterdTodos);
   } catch (error) {
     res.status(500).send(error.message);
     console.log(error.message);
   }
 });
 
-router.post('/', async (req, res) => {
+router.post('/', auth, async (req, res) => {
   const schema = Joi.object({
     name: Joi.string().min(3).max(200).required(),
     author: Joi.string().min(3).max(30),
@@ -47,7 +48,7 @@ router.post('/', async (req, res) => {
   }
 });
 
-router.put('/:id', async (req, res) => {
+router.put('/:id', auth, async (req, res) => {
   const schema = Joi.object({
     name: Joi.string().min(3).max(200).required(),
     author: Joi.string().min(3).max(30),
@@ -63,6 +64,8 @@ router.put('/:id', async (req, res) => {
     const todo = await Todo.findById(req.params.id);
 
     if (!todo) return res.status(404).send('Todo not found...');
+    if (todo.uid !== req.user._id)
+      return res.status(401).send('Todo update failed. Not authorized...');
 
     const { name, author, isComplete, date, uid } = req.body;
 
@@ -84,11 +87,14 @@ router.put('/:id', async (req, res) => {
   }
 });
 
-router.patch('/:id', async (req, res) => {
+router.patch('/:id', auth, async (req, res) => {
   try {
     const todo = await Todo.findById(req.params.id);
-
     if (!todo) return res.status(404).send('Todo not found...');
+    if (todo.uid !== req.user._id)
+      return res
+        .status(401)
+        .send('Todo check/uncheck failed. Not authorized...');
 
     const updateTodo = await Todo.findByIdAndUpdate(
       req.params.id,
@@ -105,11 +111,13 @@ router.patch('/:id', async (req, res) => {
   }
 });
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', auth, async (req, res) => {
   try {
     const todo = await Todo.findById(req.params.id);
 
     if (!todo) return res.status(404).send('Todo not found...');
+    if (todo.uid !== req.user._id)
+      return res.status(401).send('Todo deletion failed. Not authorized...');
 
     const deleteTodo = await Todo.findByIdAndDelete(req.params.id);
